feat: allow node configuration (#7)
Currently the cluster is hardcoded to have one node with 2000 MB RAM.
Some users might want to configure this. In particular, the free GitHub
runner 16 GB of RAM, so it's possible to allocate more memory to these
nodes.

I tried to allow for multi-node configuration but I think this is
difficult because it would require some extra containers in order to
give each node a separate hostname. So my quick fix was just to add a
`memory` and `cpus` input that let you configure the specs of the one
node in the cluster.

Here is an example run using my fork of this action:
https://github.com/multimeric/SlurmActionTest/actions/runs/12345117113/job/34448676056
You can see how the memory is reported as 3 GB and so is working
correctly.

---------

Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>